continue type agnostic refactor
